Rename drawing to recording
