+++ src/components/draw/BrushThicknessPickerContainer.js
@@ -5,11 +5,11 @@ import BrushThicknessPicker from './BrushThicknessPicker'
 import { setThickness, increaseThickness, decreaseThickness } from '../../actions/thickness'
 
 function mapStateToProps (state, ownProps) {
-  const thicknesses = state.drawing.thicknesses
-  const currentColor = (state.color[state.drawing.currentColor] || {}).hex
+  const thicknesses = state.recording.thicknesses
+  const currentColor = (state.color[state.recording.currentColor] || {}).hex
 
   return {
-    thicknessId: state.drawing.brushThickness,
+    thicknessId: state.recording.brushThickness,
     currentColor,
     thicknesses
   }
