+++ src/reducers/recording.js
